EG-4 Energy Revolver Port (#204)
* EG-4 Energy Revolver (#653)

* Energy Revolver

* name change

* JohnStation sprite mald pr

:pensive:

* damage rebalance

ending up with 40+ blunt damage on a heat weapon is alittle much.

6:12 changed to 3:14

* added holster sprites.

add sprites for hoslter the EG-4 on your back and suit storage

* newline-ing

maybe this will fix the check

* Cherry Pick Cleanup

Moved Projectiles.yml from DeltaV folder to Goobfolder, removed out of scope configurations

* Exped Implementation

Made it spawnable alongside the Advanced laser in t3 gun loot. Also, made it NFSD contraband.

* NFSD Uplink

Added to the NFSD Uplink and gave it a cost.

---------

Co-authored-by: Icepick <122653407+Icepicked@users.noreply.github.com>